requirements.txt
Flask==3.0.3
reportlab==4.2.2
python-docx==1.1.2
python-dotenv==1.0.1
Werkzeug==3.0.3
Pillow==10.4.0
openai==1.40.6